+++ lib/rspec_oscal_formatter/create_assessment_plan.rb
@@ -1,51 +1,43 @@
 # frozen_string_literal: true
 
-require 'date'
-require 'oscal'
+require "date"
+require "oscal"
+require "json"
 
-module RSpec
-  module RSpecOscalFormatter
-    # Create an assessment plan from the metadata and template
-    class CreateAssessmentPlan
-      attr_reader :assessment_plan
+module RSpecOscalFormatter
+  # Create an assessment plan from the metadata and template
+  class CreateAssessmentPlan
+    include CommonAssemblies
 
-      def initialize(metadata)
-        @assessment_plan =
-          Oscal::AssessmentPlan::AssessmentPlan.new(
-            {
-              uuid: metadata.assessment_plan_uuid,
-              metadata: build_ap_metadata_block(metadata),
-              import_ssp: { href: './assessment_plan.json' },
-              reviewed_controls: make_reviewed_controls(metadata)
-            }
-          )
-      end
+    def initialize(examples)
+      @examples = examples
+    end
 
-      def build_ap_metadata_block(metadata)
-        {
-          title: "Automated Testing Plan for login.gov. It #{metadata.description}",
-          last_modified: DateTime.now.iso8601,
-          version: DateTime.now.iso8601,
-          oscal_version: '1.1.2'
-        }
-      end
+    def assessment_plan
+      @assessment_plan ||=
+        Oscal::AssessmentPlan::AssessmentPlan.new(
+          {
+            uuid: RSpecOscalFormatter.configuration.assessment_plan_uuid,
+            metadata: build_ap_metadata_block,
+            import_ssp: {href: RSpecOscalFormatter.configuration.ssp_filename},
+            reviewed_controls: build_reviewed_controls
+          }
+        )
+    end
 
-      def make_reviewed_controls(metadata)
-        {
-          control_selections: [
-            include_controls: [
-              {
-                control_id: metadata.control_id,
-                statement_ids: [metadata.statement_id]
-              }
-            ]
-          ]
-        }
-      end
+    def to_json(*_args)
+      JSON.pretty_generate({"assessment-plan": assessment_plan})
+    end
 
-      def to_json(*_args)
-        assessment_plan.to_json
-      end
+    private
+
+    def build_ap_metadata_block
+      {
+        title: RSpecOscalFormatter.configuration.plan_title,
+        last_modified: DateTime.now.iso8601,
+        version: DateTime.now.iso8601,
+        oscal_version: "1.1.2"
+      }
     end
   end
 end