+++ spec/spec_helper.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require 'rspec_oscal_formatter'
+require "rspec_oscal_formatter"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = '.rspec_status'
+  config.example_status_persistence_file_path = ".rspec_status"
 
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
@@ -14,6 +14,6 @@ RSpec.configure do |config|
   end
 
   config.define_derived_metadata(file_path: %r{spec/oscal}) do |meta|
-    meta[:output_directory] = '/tmp/oscal_output'
+    meta[:output_directory] = "/tmp/oscal_output"
   end
 end