+++ test.rb
@@ -1,8 +1,0 @@
-class TestAttr
-  attr_reader :test1, :test2
-
-  def initialize
-    @test1 = '1'
-    @test2 = '2'
-  end
-end