+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+- moves some configuration to the gem level, rather than duplicating in test metadata
+- allows generating assessment plans and results separately
+- collates all test results into a single file
+- adds descriptions with test contents and location tags to easily find source of truth
+
 ## [0.1.0] - 2024-05-29
 
 - Initial release